Smoothing for thin dungeons; mark no-show chars
Characters in thin dungeons that shouldn't be shown because they're
part-way between occupy-able cells, and should be treated as floors by
existing algorithms like FOV and DijkstraMap, are changed to '\u0006',
an otherwise-unused control character, so they can be handled
separately. They are converted to spaces when printed as a string, which
should ease debugging somewhat.