Initial thin wall display code; **big RNG fix!**
There was a subtle bug in randomPortion, which is called in numerous
places, which led to (among other things, almost certainly) really bad
pathfinding in DijkstraMap. This is fixed, as is an annoying visual bug
where the animation count could get stuck above 0 (again) and games that
relied on the animation count to know when to allow input (including
EverythingDemo) would freeze when this happened. Also, StarterKit is
going to have some convenience factory methods that will make the
boilerplate of certain kinds of display easier, I think, and since thin
wall display seems like it will be somewhat finicky, this should be
useful.