Initial work on Adjacency
This new class is intended for classes where the task fo finding the
neighbors to a given cell is not as straightforward as the current
"Manhattan or Chebyshev distance 1, and done" approach. This includes
games with portals or wrapping at parts of the map, games where edges
between cells are significant (like thin-wall maps), etc. This should be
used in altered copies of DijkstraMap and FOV, at least, fairly soon.

Also, DungeonGenerator and its variants, plus DijkstraMap, are marked as
Serializable now.