Very early version of CustomDijkstraMap
This is mostly a subset of DijkstraMap that allows custom adjacency
rules to be specified for portals, wraparound maps, thin-wall maps,
games with rotation as a kind of movement, and so on. Not all the
features are currently possible; thin-wall maps still need a small
amount of extra code and rotation needs a fair bit more. Currently it
has the most-specialized features of DijkstraMap removed and the code
for attack pathfinding commented out, waiting on a custom version of LOS
to be implemented.