Swap out Lightning hash implementation
The existing one had some critical flaws, such as always giving a hash
of 1 to two-element arrays with the same element twice. The new one is
tested a bit more and seems much more durable, and also faster. The
technique it uses is documented in the nested class JavaDocs, and a
HashQualityTest has been added.