Finishing step in Lightning improves quality
It's improved enough that it seems on-par with SipHash in a new test of
visual artifacts produced when a hash of an int[2] array (representing
an (x,y) point) is used as a 24-bit RGB color for the pixel at that
point. The default Arrays.hashCode method performs shockingly poorly at
this with only black and shades of blue generated, FNV-1A has noticeable
artifacts but many colors, and SipHash has only a handful of artifacts
as vertical lines or folds. Lightning, with the new finishing step,
looks almost identical to SipHash except for a few horizontal folds at
predictable locations (mainly high multiples of powers of 2, it seems),
and even those are tiny vertically. It doesn't seem to slow it down too
much, either, from 17 to 19 ms over a million 16-element arrays
including their production time.