Slight speedup/quality improvement in Lightning
Also, updates to the docs, including visual previews of the hashes for
x,y points as the colors for the pixels at those points. This info was
produced by the last commit and updated for the latest change to
Lightning. This actually has more hash collisions on the sample data
than pre-finalization, but still less than the other tested hashes, and
its worst-case collision count/frequency is significantly improved.