Lightning should be done at some point...
There was a potential issue with the finalizing step that wasn't visible
with 32-bit hashing but was in 64-bit hashing. After some weird
tinkering, Long.rotateLeft (which is intrinsic on some JVMs, making it
much faster) was found to yield the least predictable hashes for all
tested areas of bits, along with some other not-very-easy choices of
constants and operations to determine the distance of rotation and what
to actually rotate. The final result has very slightly lower speed, and
is likely to not have a perceptible difference. A bug was fixed in
SipHash's hash function for the overloads that take a char array, a
starting index, and an ending index, where the indices were partly
ignored. CrossHash (FNV algorithm) also got that same overload `(char[],
int, int)` added for a 32-bit hash instead of only the 64-bit hash it
had before.

HashVisualizer has a lot more options that can be cycled through with
the Enter key. Checking the source may be needed to tell what you're
looking at for now, but FNV and Arrays.hashCode both have distinctive
flaws.