Add ThunderRNG (FAST), StringKit gives more info
Also, some small changes in SobolQRNG and a tiny change in CrossHash to
make DefaultHasher act like the other IHasher implementations.
ThunderRNG needs more work to determine how usable it will be, but it
currently seems to be random enough to be indistinguishable in bit
frequency from other good RNGs like LightRNG. It's also the first
RandomnessSource that I've found that generates a billion random longs
of (probably) sufficient quality in under a second on this laptop, with
second-place close between LightRNG, XoRoShiRo128+, or libGDX's
XorShift128+ implementation.