Better Thunder beats LCGs on speed, quality
Thunder now doesn't seem to fall into a known category of RNG, but for
some reason the JVM seems particularly able to optimize the current
approach, which uses only an addition with assignment, a bitwise signed
shift, a XOR, and a multiplication. It has less quality problems than
the linear congruential generator tested with the same tool, and is
almost twice as fast (it is unclear why such a simple LCG is this slow
on the JVM, though... The last commented line in ThunderRNG.nextLong is
the LCG, and it benchmarked at almost equivalent speed to LightRNG, both
within 5% of 1250ms for the same data, while ThunderRNG currently takes
about 715 ms).