Add Storm hash family, tweaks to Lightning
Lightning had a slight flaw, possibly, where the least significant bit
was significantly less likely to be set, though a bitwise rotation after
the flaw occured distributed it across all 64 bits.

Storm is based on Lightning and is about half the speed, but allows a
64-bit salt to modify the hashing function in a wide amount of ways
(halfway on its way to Sip, which allows 128 bits of salt and is about a
quarter the speed of Lightning). The idea is to maybe replace Sip at
some point with something that has similar quality, and Storm seems only
slightly worse on visual quality and significantly better on collision
rates. Sip is mainly a lot of code for such a similar result to
Lightning and Storm, and it is just not fast enough to warrant its
continued inclusion, in my opinion.

String hashing functions in Lightning and Storm now don't create a
temporary char[] and just read chars from the String, to avoid creating
garbage.