ThunderRNG gets improvements to period
Its speed should be somewhat behind where it was, but the
lower-than-normal period on the old version could have been a problem
(and for less-significant bits, like from calling next(2), the period
was low enough that with the right count of numbers generated per frame,
every frame would get identical "random" numbers). Several fixes help
here; ThunderRNG tracks two separate state parts and increments one by
the a portion of the other. This behavior helps keep the period of all
the bits higher. ThunderRNG is heavier-weight now, with two longs, but
it is faster than the other RandomnessSources, as the latest set of
results in RNGBenchmark show.

There's also some tinkering in HashVisualizer; press s to switch to the
Storm visualizations, then s again to switch to "artistic
visualizations" where I'm trying to make some kind of map or other
semi-random shape from rather-random data.