Fix some JavaDoc warnings, speed up Storm Hash
Storm also comes with a bunch of predefined hashing functions that can
be used conveniently, and it can be initialized with a String (that gets
hashed by Lightning to get a long). The main change to Storm was to
remove essentially all of the old code and replace it with Lightning's
algorithm, plus a small step at the end that mixes in a value calculated
when the Storm object was constructed. The visual quality and number of
collisions in test data both seem fine and the speed is better than
before.

The JavaDoc warnings were mentioned in an issue about a year ago, and
most had been fixed either manually by @smelc or by a Java-8-specific
profile added some time ago that prevents overzealous doc linting by the
JDK. Only a few warnings are left, and mostly in huge files like
SquidPanel.