Manual merge of #131 , various other things
This keeps the existing API for tint and slide in SquidPanel but adds
@smelc 's enhanced versions that allow a number of Runnable objects to
be passed and run after the animation completes. This may have some
issues in smelc's code if his version from the PR and the older code
without varargs have non-equivalent behavior when no Runnables are
passed; if the empty-vararg case from the PR is equivalent to the
no-Runnable-arguments version in this code, then there should be no
mismatch. It also includes the ColoredString and Markup fix that smelc
had in PR 131.

Other than that, there's some minor cleanup in PermutedRNG, some recent
benchmark results, and some prep for an update to libGDX 1.9.4 (it
currently breaks font scaling, and changes may be needed to be
compatible).