Fix long-standing statistical issue with some RNGs
tl;dr version: This changes some RandomnessSource implementations to fix
a quality bug. LightRNG is essentially unaffected, so most of SquidLib
isn't affected either, but LongPeriodRNG did change its core algorithm
to fix an old bug, and that may have some effect.

The only RandomnessSource of this group that is commonly used in
SquidLib and has its core algorithm heavily affected is LongPeriodRNG,
which, like XorRNG and the same-algorithm ChaosRNG, used the
sign-extending right shift operator instead of the correct *unsigned*
right shift operator in a few places, and this caused a drastically
higher likelihood that the sign bit would change with each generated
number, to about 2/3 of generated pairs having different signs instead
of the correct 1/2 on average. The only other places where this change
matters are in various nextBytes() methods, where there seems to have
been a bias in every 64th bit generated (I think it was always 0); now
the bias should be gone. ChaosRNG also marks all of its fields as
transient now, since it shouldn't make its state visible when
serialized.