Fix bug in XoRoRNG, improve Storm, ThunderRNG
XoRoRNG had an issue with initialization that probably caused the second
number generated to be 0 every time that certain common initializing
methods were used. This is fixed. Storm now does a bit better of a job
at concealing the behavior of the underlying, main algorithm of
Lightning that it is built on, by using the salt differently. ThunderRNG
has a better period for the bottom 2 bits of each long, up from about
2^18 to at least 2^20 for the minimum 2 bits and some higher number (at
least 2^41, possibly 2^127 but not higher, and I decided not to wait for
a brute-force test to complete after a few weeks or months to confirm
the actual period) for the full 64 bits.