Optimizations to MultiSpill, fixes in GwtCompat
MultiSpill now involves no foreach iterations and less checks for each
added cell, which should help performance. A small possibly-breaking
change in its API requires an OrderedMap instead of a general Map
because the ordering actually matters when the keyset is put into an
ArrayList with specific meaning for each index, and because OrderedMap
allows random access of values. MultiSpill is being evaluated for usage
in worldmap generation, possibly in conjunction with Perlin Noise a la
MetaMapGenerator.

In GwtCompatibility, range(int start, int end) was broken, but since it
was never used in SquidLib it stayed hidden for a while. A similar
method was added for making character ranges (but charSpan(char, char)
is inclusive on both ends, to match the behavior of regex char classes
and the normal terminology of "A to Z"), which can be useful when
debugging among other things.