Various fixes suggested by @smelc
DungeonGenerator and related classes (like SectionDungeonGenerator) now
check that a door isn't placed in a pointless position, such as blocking
passage through one cell when a way to walk around the door is clearly
visible and almost as short a path. This uses a Dijkstra-based technique
recommended by @derrickcreamer , thank you! TextScreen is marked as
deprecated because it uses GroupCombinedPanel. More stuff in
DungeonGenerator uses GreasedRegion for speed reasons. GreasedRegion and
DungeonUtility have some more features.