Add better noise to Merlin using GreasedRegion
GreasedRegion now has a constructor that takes a RandomnessSource to
quickly fill it with random values, and MerlinNoise has a way to use
that white noise with GreasedRegion's retract() and expand() methods to
create a series of layers of diamond-shaped regions. The new bitSum
method in GreasedRegion lets us treat a cell in each element of an array
of GreasedRegions as an int, with each GreasedRegion contributing one
bit of the int. That is used by MerlinNoise to precalculate nice noise
from the layered diamonds made earlier.