Add Object[] hashing for Storm and Lightning
This delegates to the Object's hashCode() method if defined, or whatever
hashCode its superclass defines (possibly the one on Object, which is
always present). Some classes that didn't have hashCode or equals()
methods have them now; it's just better style to define them. If
Lightning or Storm defines hash() or hash64() for arrays of a reference
type more specific than Object, then the more specific hashing function
will be used.