Starting work on serialization in -extra module
This prevents a dependency on libGDX for squidlib-util, but allows
functionality that needs utility code from libGDX, like Preferences and
its Json classes, for games that already use libGDX either for
text-based graphics using the display module or sprite-based graphics
using libGDX without the display module. Much work was needed to get
serialization working correctly for Maps with non-String keys (a
limitation normally in place on libGDX Maps, but it works now, and with
BlazingChain as a small dependency for th -extra module only, saved info
in preferences are shrunk from 50K to 5K for a bunch of test data, and
it reads it back in correctly.