Update RegExodus (don't use unavailable snapshot)
RegExodus had a long-standing bug in the toString method for Pattern. I
say long-standing, and I mean about 15 years, though for most of that
time it was abandoned. But it still made debugging slow and could crash
during serialization of Patterns, and the latest update finally seems to
fix it. I was testing on a -SNAPSHOT version built locally, but now that
0.1.7 is on Maven Central, we can depend on it here. SquidStorage also
has a much more compact representation of `char[][]` values, and should
have nice short type tags in many places.