GWT compatibility; huge FLG serialization boost
FakeLanguageGen now serializes to a tiny fraction of the size it
formerly took to store a language, and it does this for most kinds of
languages that are practically used in SquidLib (not custom ones, but
mixes of multiple languages made with the new methods for more sensible
ratios between languages, and languages with modifiers applied, will
usually have the right info).

Several other changes help GWT compatibility, and the reflection cache
needs to be investigated to see how small we can make it.