Minor improvement to Thesaurus, trying JitPack fix
There was something wrong with squilib-extra's pom.xml file, which
prevented JitPack from building correctly, but I think it's fixed now.
Thesaurus now lets you retrieve the random languages it generates for
nation names, and although it doesn't hold on to them for long, it
allows users to generate more text in the same style as a nation's name.