Important changes that make -extra saving work
BlazingChain, the recently-added-to-squidlib-extra dependency that
allows compressing JSON and other text, needed a minor configuration
update to work on GWT. SquidStorage needed some significant changes to
work in more general cases than the unit test it had been checked
against, but now it is confirmed to work on desktop and GWT (the demo
should be in tommyettinger/SquidLib-Demos soon). The reflection cache
was way too big before, and it's now using a more reasonable amount of
cache space.