Rotating direction markers work!
This was way more difficult than I thought it would be, since libGDX
BitmapFonts don't allow themselves to be rotated at all. I got the code
to extract the boundaries for the single char '^', pull that into a
TextureRegion, and use that to create Images where needed. Images can be
rotated, even if BitmapFonts and Labels can't. The example is in
SquidLayersTest.