Many small changes
Removed encoding methods in StringKit (they didn't help reduce the size
of textual representations of numbers). Improvements to speed for
MerlinNoise, especially in 3D. Re-added some methods that slipped by me
when I added them to the stand-alone RNG project, Sarong, but are also
useful here (in GwtCompatibility, including invertOrdering for the
orderings that RNG can produce, and reorder for ArrayLists as well as
the existing OrderedMap/Set stuff). RNG now has a shuffleInPlace method
that is GWT-compatible.

This commit seems huge because it removes timestamps on all the docs,
but future commits shouldn't change every javadoc HTML page just to
update timestamps, since I set some configuration in Maven to avoid that
from now on.