RoomFinder assigns all its fields now.
Before, there was an issue (exposed in SectionDungeonGenerator when a
map char[][] was passed with no environment int[][]) where allCaves
would not be given a value if a rooms-and-corridors map was being
generated, but SDG needs allCaves present for its lake placement. This
was a pretty simple fix, and it exposed a few places where
GreasedRegion.writeInts() was being called without using the result; the
result is now used.