Fix FakeLanguageGen when using char limit
Before, sentences made with one method could have the last word of the
sentence, the ending punctuation, and then a repeat of the last word due
to a bug in one overload of the sentence() method that limited maximum
chars in the sentence. This has been fixed, and should be most
noticeable in BasicDemo.