Need for Speed, Hashing Edition
CrossHash.Falcon beats Arrays.hashCode() on 32-bit hashing speed, which
is impressive considering Arrays.hashCode is incredibly minimal and
collides often, while Falcon doesn't collide that differently from
Lightning, Storm, and FNV-1a. Visual tests still should be run on it.

ThunderRNG now goes through one (partial) generation of a random number
when a 64-bit seed is given to it, since before there was a less-random
quality to the first number generated, and later numbers were better.
This changes numbers generated with ThunderRNG instances seeded with the
single long parameter, but not the two-long-parameter constructor.

Slight tweaks to FakeLanguageGen based on issues found in Gabber, the
spinoff project that has just SquidLib's language-generation code and
not much else.

DijkstraBenchmark now tests against gdx-ai's good-but-clumsy-to-use A*
implementation, which outperforms DijkstraMap but not too harshly
(gdx-ai is between 3x and 4x faster but can only pathfind to one goal at
a time).