Change GapShuffler. Better Falcon hash.
GapShuffler now doesn't need to produce new GapIterators, since it is
both an Iterator and an Iterable now. This changes the API slightly,
from getNext() to next(), so it matches the Iterator interface.

It seems HashVisualizer is useful yet again for finding problems in
hashes; Falcon was and still is fast, but before it had rather severe
issues in its visual hashes, meaning similar inputs would weight certain
elements in the hashed arrays much more heavily than others, and the
full range of values didn't reliably appear. This has been fixed with a
very small finalization step taken from the approach used in FastUtil
(which got it from Koloboke), where instead of the result being returned
directly, it is multiplied by the golden ratio as an int or long, then
XORed with itself shifted over by 16 or 32 bits. The constants used in
the 32-bit Falcon hash also changed, after much experimentation, but the
64-bit hash didn't need any further tweaks.