More minor changes to Falcon hashing.
It now has a better finalization step that seems to eliminate the
"banding" or "dappling" artifact in HashVisualizer when testing Falcon
with one-element arrays, and that means it's doing a better job at
dispersing the input bits. It also gets quite a bit fewer collisions in
HashQualityTest, and its speed seems indistinguishable in HashBenchmark,
so win-win-win.