Added Arrangement, an object to int mapping
Arrangement isn't especially useful standing alone, but it becomes much
more useful when you realize it can be used to flexibly access any
ordered collection with any type that can be hashed (you get an int from
Arrangement.getInt(), then look up at item with getAt() or keyAt() in
another collection), plus you can look up elements in Arrangements in
reverse as well. I've wanted to have a class like this in SquidLib or
projects using it on a few occasions, and I figured it might as well be
time to add it.