Efficiently serialize Arrangement
SquidStorage now correctly avoids serializing the value array in
Arrangement, since it is assigned by inserting elements into the data
structure and none of the constructors use it.

FakeLanguageGen was essentially using its own older style of Arrangement
with an OrderedMap that had Integer keys. But because it never needed to
look up the Integer associated with a key, it wasn't really a Map, so it
doesn't need Arrangement, and now it does fine with an OrderedSet.