Trying to understand noise optimization
I'm preparing the HashVisualizer code for testing a new noise function
someone offered to write, and I wanted to better understand what affects
the performance of MerlinNoise relative to our implementation of
PerlinNoise. I found some artifacts present in MerlinNoise, figured out
a simple way to avoid recalculating a sine and a non-integer exponent
for every noise call (lookup tables), and simplified the rawNoise
functions for the new code. This makes MerlinNoise somewhat faster than
PerlinNoise in both the 2D and 3D cases, although the visual properties
of the noise are very different. The old MerlinNoise code is still
there, for now, but I may consolidate noise2D with noise2D_alt,
replacing the former with the latter, and do the same with the 3D
versions.