Add SquidIterators.Linear and 4 directions for it
Credit for most of this goes to @smelc , I just generalized his code for
iterating left-to-right so you can give a Direction in the constructor,
and made Up, Down, Left, and Right as simple subclasses. I may have
screwed something up in here, but I'd like to build a certain project
that needs this.