Bring in Messaging from related Gabber project
This is a handy class for writing messages that the user should see, and
upholding the DRY principle while doing so. It handles conjugation of
English verbs and pronouns in a fairly robust way, adjusts nouns for
pluralization, and is extensible to handle irregular words. It can
handle just about any gender preference need, possibly by delegating to
lib-using code that can post-process its output, but usually doesn't
need any hard work. It currently is limited to one "user" and one
"target", both of which can be singular or plural, but can't deal with
the ambiguity of multiple targets. I hope it's useful! There's examples
in the javadocs.