Mid-quality version of Norse in FLG
This should help bridge the gap between what is supported by
WeightedLetterNamegen and what is supported by FakeLanguageGen.
Thankfully, Icelandic is very close to Norse linguistically, so I didn't
need to figure out a romanization scheme for ancient runes, and I can
use a modern and well-supported-by-fonts set of characters.