Add PintRNG, a RandomnessSource meant for GWT
PintRNG uses ints almost exclusively (pint-size numbers compared to
longs), so avoids using slower emulated longs on GWT. It is not terribly
fast on other platforms, about 2/3 to 1/2 the rate of LightRNG on
desktop when generating ints and 1/4 to 1/5 the rate when generating
longs. There are places where performance could improve. Its statistical
quality is uncertain, but is probably good enough for a game in the
browser (flawed RNGs have been used by websites frequently, notably
including Microsoft's former webpage to allow choosing an alternate
browser for Windows in the EU; this doesn't appear to be as flawed at
first glance, but it could be).