Add Filters.DistinctRedGreenFilter for...
... better supporting accessibility for players with one of the two most
common kinds of color-blindness, while having minimal effect on players
without color-blindness. Try it in EverythingDemo by pressing 'f' to
cycle between filters. Mainly, it changes the brightness of red and
green colors so greens are more vivid to those with "normal" vision
while reds are slightly darker and more-gray. For people with
deuteranopia, what is normally green should appear brighter, while what
is normally red should appear darker; for people with protanopia, this
change is the same but the effect is much stronger.