Add VanDerCorputQRNG, improve some javadocs
The first implements the van der Corput sequence, which is useful as
part of generating random more-uniformly-spaced points with the Halton
sequence, as well as some other uses. SquidInput has probably been
giving JavaDoc warnings this whole time due to unescaped '{' and '}'
characters in the docstrings, and if it was giving warnings, they're now
fixed.