Add iterator for GreasedRegion, halton in VDC
The static halton() methods in VanDerCorputQRNG could be in any class,
really; they don't use any specific data to that class or create any
instances of it, but the documentation for how van der Corput and Halton
sequences work is in there, so I guess it fits.

GreasedRegion now also implements Iterable<Coord>, though this needs to
be tested still. It is implemented similarly to the existing first()
method, and the first item during iteration should be the same produced
by first().