Bring CustomDijkstra API closer to DijkstraMap
So far, this is just the partial scanning methods added in this commit,
but that's enough to benchmark CustomDijkstraMap (CDM) on equal footing
to DijkstraMap. Rather surprisingly, CustomDijkstraMap is faster when
using an equivalent Adjacency to what DijkstraMap uses, and is better
than twice as fast at finding very short paths with a partial scan (a
common use in roguelikes, potentially). This was not exactly expected,
since CDM is a lot more flexible than DijkstraMap and that could have
been a slowing factor. It is good though! CDM still needs some support
classes to be able to do all the same things that DijkstraMap does.