AStarExperiment in -performance shows promise
Approximately 10x faster than the current AStarSearch, but still about
16x slower than the one in gdx-ai. Could be used as a drop-in
replacement.

This commit also fixes a bug in OrderedSet, OrderedMap, and Arrangement
where removing the only item in one of those collections would not
remove that item's position in the ordering. Proceeding to add items and
get them with .getAt() would then have strange and undefined behavior,
usually depending on the default return value. This appears to be fixed.