Replace AStarSearch with optimized version
The algorithm is almost the same but doesn't suffer from recursive call
count explosions, and so is much faster. It's faster than DijkstraMap on
short paths, and slower on long paths. There may be bugs in this
version, but there were more serious ones in the old one (on long paths,
it could overflow the stack).