Various small tweaks; PerlinNoise (2D) change
There were some issues with PerlinNoise when producing 2D noise, most
noticeable when only one octave was generated. Artifacts tended to show
up on 45-degree-angle lines, and it strangely used 12 permutations with
some duplicates (the same as it used in 3D, but discarding z). This goes
against Perlin's recommendations, but the technique we use only does a
little less of that. Our gradents for 2D are all on the unit circle, and
are based by starting at the point at 0 radians and advancing phi
radians for each subsequent point, where phi is the golden ratio. This
definitely reduces how noticeable artifacts are.

Also changed are Coord's hashCode() method, which was rather flawed
before, some simple implementations of equals() and hashCode() for RNG,
and some hash visualization stuff. IColorCenter.Skeleton has a new
method that may be handy in debugging to see how many colors you're
actually using.