Improve fitting used by TextCellFactory
This was far more difficult than it should have been. This is already
slightly noticeable as a better fit because ThinWallDemo doesn't have as
much light shine through walls (about half as much, and this is mainly
dependent on the font's box drawing characters). I don't know how much
effect it will have on other fonts, but it should generally be an
improvement when adding new fonts that use a standard set of
well-supported characters from Latin, Greek and Cyrillic.