Readability improved with NatLangCipher
Now NaturalLanguageCipher won't generate certain hard-to-pronounce
blends of letters when the suffix is a poor match for the base word. In
Latin-script text, blends such as the endings "iy" or "uy" or the
combination anywhere "qq" will be changed to "y", "y", and "q"
respectively. This isn't perfect, but it helps reduce the frequency of
very strange words.