Some cleanup in GreasedRegion, better VDC points
GreasedRegion somehow used the constructor for Coord instead of
Coord.get(), which meant some of  the Coords it returned (it didn't
store any internally) were not pooled. This is fixed. There's now an
alternate way to get deterministic but random-seeming points that avoid
overlapping too often using VanDerCorputQRNG (or, more likely, the code
that wraps it in GreasedRegion, GreasedRegion.quasiRandomSeparated() ).
GreasedRegion.nth() provides useful functionality that CoordPacker had
but until now GreasedRegion lacked, plus we have
GreasedRegion.atFraction() to get the Coord some percentage or fraction
of the distance through the iteration order.