Rework tilesets so they use GreasedRegion inside
This does need to be checked on GWT if anyone uses Tileset-based dungeon
generation there, since GreasedRegion is compatible but may be slower by
some unknown margin due to long emulation. But, that will be less
noticeable when we can update to GWT 2.8.0 with the next release of
LibGDX. The advantage to this method is mostly that it's cleaner
internally, with less transposing needed and less individual writes to
array values when writing a whole column can be done instead. There's a
lot of other small changes/additions here, like several useful additions
to StringKit and some small additions to GreasedRegion (mostly for
generating code).