Vital fix to dungeon gen; burst() and summon()
The fix to dungeon gen is needed because the tests wouldn't pass due to
runtime crashes whenever dungeons with more than 64 height were made.
There was a bug regarding GreasedRegion's newly-added constructor where
it didn't properly resize the input data and could get array bounds
violations. That's fixed now and the tests pass.

SquidPanel has a summon() method that takes a delay parameter now, and
summon() is made more convenient by the various overloads of burst(),
which create a splashing or exploding effect.