Improvements to noise, OrderedMap/Set fixes
OrderedMap and OrderedSet had some rather serious issues when using the
additional API (not in LinkedHashMap/Set) for altering a key in-place;
using alter() could have caused the size to increase but pad values with
null keys and/or values, causing mysterious NPEs. When updating
IconsTest to use OrderedMap instead of SpatialMap, I encountered this
bug, found its cause (bad alter() implementation), and fixed it (by
adding functionality to add or put at a given point in the ordering,
which is most of what alter() needs to do after removing a key).
OrderedMap and OrderedSet both passed the OpenJDK test suite this whole
time, but since this is an extra API, it wasn't tested.

WhirlingNoise has some nice improvements, and should be fine even
unscaled at integer coordinates, which our PerlinNoise needs to scale to
accomplish.