Bugfix in Dijkstra; GreasedRegion is a Collection
DijkstraMap pathfinding methods didn't adequately check that they were
checking in-bounds for their 2D array, so that could cause issues
(thanks,  @zacharycarter for finding this).

API CHANGE (minor):
GreasedRegion now implements Collection, so the clear()method it had
before, which returned a GreasedRegion, now returns void. The old
behavior is still present in the empty() method, which clears and
returns the cleared GreasedRegion for chaining. Also added are several
other Collection methods, and MultiSpill now takes a Collection<Coord>
instead of a Set<Coord> in start() , so you can pass a GreasedRegion to
it and save heavily on memory usage if you already had a GreasedRegion
(this is the case in SpillWorldMap and it seems to speed it up quite a
bit).