Fix iterator in GreasedRegion, DM can take Colls
And for @zacharycarter, DijkstraRiverTest may be useful to run; it uses
both of the other added features/fixes, plus DijkstraMap now sets a
boolean, cutShort, to true after pathfinding if it couldn't reach a goal
(false if it found a goal). This is used to place lakes in the test.