New (hard to write!) method in RNG, Greased
This method in RNG, approximateBits(), produces a random long with a
strong statistical bias toward generating numbers with a specified bit
count (AKA population count, Hamming weight, popcnt, etc.). It is not
guaranteed to generate numbers with that count except when 0 and 64 are
requested, but this can be useful for when bits themselves are data.
It's used in GreasedRegion with the constructors that take RNG
parameters now optionally also taking a double fraction for how much of
the map should be "on", implemented with RNG.approximateBits()
internally. That's used in DijkstraRiverTest to show how to change the
"wiggle level" of rivers.