GreasedRegion stuff for removing nearby cells
Meant for the river-drawing code that @zacharycarter is making,
GreasedRegion.disperseRandom() should be a good way of modifying a
randomly-filled GreasedRegion so no cells are too close by. The random
fill given to the constructor or refill() can be much higher if
disperse(), disperseRandom(), or disperse8way() is called after, since
the first two set about half the bits to 0 and the last sets about 3/4
to 0.