Major improvements to WhirlingNoise
WhirlingNoise is sort of an alternate approach to the existing simplex
noise algorithm as implemented in PerlinNoise. Only now, WhirlingNoise
is a fair bit faster, about 20% to 25%, while having better statistical
qualities that make it less repetitive. It uses part of the algorithm
used by PintRNG (and PermutedRNG) to quickly generate wildly different
ints from each input int, and this replaces a series of lookups in a
256-element array with fixed contents.

WaffleFill is on hold for a little while, as I attempt to debug it more
effectively.