FontTest hurts my brain, but it works.
There were several bugs in the last commit relating to DefaultResources,
and all seem to be fixed now. Calling getStretchableSciFiFont() would
cache the produced TextCellFactory in the wrong variable, overlapping
with the stretchable typewriter one, and that was only noticeable in the
new FontTest that tries lots of fonts for display purposes. Not all the
TextCellFactory variables were being properly disposed in the pause()
method of DefaultReources, either, so that's fixed as well. Threee
fonts, generated by https://github.com/tommyettinger/Glamer , have been
updated to look a lot better and not get cut off at the top (for some of
them, this was a problem). All of the fonts that are fairly large in
size have also been optimized for size with pngout, which is a very
effective tool for lossless size reduction.