Strongly universal hash functions in CrossHash
These are in the new Chariot nested class, and implement code based on
Daniel Lemire's work (link in javadocs). Chariot is slightly faster than
Storm while having better statistical properties (it seems, anyway; they
both collide rarely on the given inputs), but fails visual testing
(similar patterns of inputs produce similar patterns of results). It
looks like it could be handy, and it allows 128 bits of salt, which
makes it usable for the stuff Sip was in here for.