Needed a better sci-fi font, Orbitron is great
Orbitron is a variable-width distance field font here, like our versions
of Gentium and Noto Sans. It has a blocky,
"interstellar-mining-corporation" feel to it, and should be useful in
menus with classes like TextPanel.

Also, the assets zip is updated, and DefaultResources has a version of
the "Orbit" font pre-configured.