Various tweaks to Permuted and Pint, better speed
Since PintRNG is specifically meant for environments where speed is
critical and 64-bit math is slow, having it get faster should be good.
The LCG-like step that advances the state on PermutedRNG and PintRNG has
been replaced with LightRNG's simpler "state += phi" approach, where phi
is the golden ratio times 2^64 (or 2^32 for PintRNG). This doesn't seem
to affect quality much, if at all, but does improve speed a lot for
PintRNG and a little for PermutedRNG.