Some tweaks to ProbabilityTable, serialization
You can now set the RNG in a ProbabilityTable after constructing it,
which should help address one recent issue... @SquidPony .
ProbabilityTable produces more tightly-condensed sorta-JSON, as can be
seen by running squidlib-extra's SquidStorageTest (which currently just
tests ProbabiliyTable serialization and deserialization). It still needs
to say what types the serialized keys are (the <T> type in
ProbabilityTable<T> allows subclasses, so it's needed). The sorta-JSON
produced by SquidStorage should in general be smaller by omitting the
class name String where it's always the same in the outermost layer of
the JSON, both for keys and values; this was done by subclassing
OrderedMap to make an always-String-keys-and-values Map,
StringStringMap, which should add only a handful of constructors, but
might add a lot more (depending on how Android and GWT count them).