Augment LockTest, MixedGenerator, more
MixedGenerator now modifies the IRoomLayout that can be passed to a
constructor, making the centers of Rooms in it match their actual
locations in the dungeon MixedGenerator will produce. LockTest now shows
where keys are and what passages require keys to pass through; this
seems accurate so far. ArrayTools now has a fast way to look up a letter
given an int, and letterAt(int) won't throw exceptions if its parameter
is negative or very large (it wraps around politely).